feat: provide custom title for each page. feat: add global error. feat: add favicon image.
