+++ src/app/error.tsx
@@ -1,4 +1,5 @@
 'use client';
+
 export default function Error({
   error,
   reset,
