+++ src/app/profile/[id]/page.tsx
@@ -1,7 +1,12 @@
 import UserPosts from '@/components/UserPosts';
 import UserProfile from '@/components/UserProfile';
 import { getConvexClient } from '@/lib/convex';
+import { Metadata } from 'next';
 import { api } from '../../../../convex/_generated/api';
+
+export const metadata: Metadata = {
+  title: 'conveX | Profile',
+};
 
 export default async function UserProfilePage({
   params,
